// client/src/features/members/member-profile/member-profile.ts
import { Component, HostListener, OnDestroy, OnInit, viewChild } from '@angular/core';
import { inject, signal } from '@angular/core';
import { ActivatedRoute } from '@angular/router';
import { Member } from '../../../type/member';
import { DatePipe } from '@angular/common';
import { MemberService } from '../../../core/service/member-service';
import { EditableMember } from '../../../type/member';
import { ViewChild } from '@angular/core';
import { ToastService } from '../../../core/service/toast-service';
import { FormsModule, NgForm } from '@angular/forms';
import { AccountService } from '../../../core/service/account-service';

@Component({
  selector: 'app-member-profile',
  imports: [DatePipe,FormsModule],
  templateUrl: './member-profile.html',
  styleUrl: './member-profile.css',
})
export class MemberProfile implements OnInit ,OnDestroy{
 
  @ViewChild('editForm') editForm?:NgForm;
  @HostListener('window:beforeunload',['$event']) notify($event:BeforeUnloadEvent){
    if(this.editForm?.dirty){
      $event.preventDefault();
    }
  }
   private accountService=inject(AccountService);
  protected memberService=inject(MemberService);
  private tost =inject(ToastService);
protected editableMember:EditableMember={
  displayName:'',
  description:'',
  city:'',
  country:'',
};



ngOnInit(): void {
  
  
  this.editableMember={
  displayName:this.memberService.member()?.displayName|| '',
  description:this.memberService.member()?.description || '',
  city:this.memberService.member()?.city || '',
  country:this.memberService.member()?.country || '',
}

}
updateProfile(){
  if(!this.memberService.member) return;
  const updateMember={...this.memberService.member(),...this.editableMember} as Member;
  this.memberService.UpdateMember(this.editableMember).subscribe({
    next:()=>{
        const currentUser= this.accountService.currentUser();
        if(currentUser && updateMember.displayName !== currentUser?.dispName){
        currentUser.dispName=updateMember.displayName;
        this.accountService.setCurrentUser(currentUser);
        }

        this.tost.success('Profile updated successfully');
        this.memberService.editMode.set(false);
        this.memberService.member.set(updateMember as Member);
        this.editForm?.reset(updateMember);

    }
  })

}
 ngOnDestroy(): void {
  if(this.memberService.editMode()){
  this.memberService.editMode.set(false);
  }
  }
}
